@@ -1,16 +1,20 @@
 {
   "name": "splayer",
-  "version": "2.0.0",
+  "version": "2.0.2",
   "description": "A minimalist music player",
   "main": "./out/main/index.js",
   "author": "imsyy",
   "home": "https://imsyy.top",
   "github": "https://github.com/imsyy/SPlayer",
   "repository": "github:imsyy/SPlayer",
+  "license": "AGPL-3.0",
+  "license-file": "LICENSE",
   "engines": {
-    "node": ">=16.16.0"
+    "node": ">=18.16.0",
+    "npm": ">=9.6.7",
+    "pnpm": ">=8.14.0"
   },
-  "packageManager": "pnpm@8.12.0",
+  "type": "module",
   "scripts": {
     "format": "prettier --write .",
     "lint": "eslint . --ext .js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts,.vue --fix",
@@ -23,8 +27,8 @@
     "build:linux": "npm run build && electron-builder --linux --config"
   },
   "dependencies": {
-    "@electron-toolkit/preload": "^2.0.0",
-    "@electron-toolkit/utils": "^2.0.1",
+    "@electron-toolkit/preload": "^3.0.0",
+    "@electron-toolkit/utils": "^3.0.0",
     "@material/material-color-utilities": "^0.2.7",
     "NeteaseCloudMusicApi": "^4.14.0",
     "axios": "^1.6.5",
@@ -33,12 +37,12 @@
     "electron-store": "^8.1.0",
     "electron-updater": "^6.1.7",
     "express": "^4.18.2",
-    "express-http-proxy": "^1.6.3",
+    "express-http-proxy": "^2.0.0",
     "howler": "^2.2.4",
     "js-cookie": "^3.0.5",
     "localforage": "^1.10.0",
-    "music-metadata": "7.13.4",
-    "node-id3": "^0.2.6",
+    "music-metadata": "7.14.0",
+    "node-taglib-sharp": "^5.2.3",
     "pinia": "^2.1.7",
     "pinia-plugin-persistedstate": "^3.2.1",
     "plyr": "^3.7.8",
@@ -49,24 +53,24 @@
   "devDependencies": {
     "@electron-toolkit/eslint-config": "^1.0.2",
     "@rushstack/eslint-patch": "^1.6.1",
-    "@vitejs/plugin-vue": "^4.6.2",
-    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vitejs/plugin-vue": "^5.0.2",
+    "@vue/eslint-config-prettier": "^9.0.0",
     "ajv": "^8.12.0",
-    "electron": "^27.2.1",
+    "electron": "^28.1.2",
     "electron-builder": "^24.9.1",
     "electron-log": "^5.0.3",
-    "electron-vite": "^1.0.29",
+    "electron-vite": "^2.0.0",
     "eslint": "^8.56.0",
     "eslint-plugin-vue": "^9.19.2",
-    "naive-ui": "^2.37.0",
+    "naive-ui": "^2.37.3",
     "prettier": "^3.1.1",
     "sass": "^1.69.7",
     "terser": "^5.26.0",
-    "unplugin-auto-import": "^0.16.7",
-    "unplugin-vue-components": "^0.25.2",
-    "vite": "^4.5.1",
+    "unplugin-auto-import": "^0.17.3",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.0.11",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-pwa": "^0.17.4",
-    "vue": "3.4.4"
+    "vue": "3.4.8"
   }
 }
