@@ -19,11 +19,20 @@ const useMusicDataStore = defineStore("musicData", {
       playSongSource: 0,
       // 当前歌曲歌词数据
       playSongLyric: {
-        lrc: [],
-        yrc: [],
-        hasTran: false,
-        hasRoma: false,
+        // 是否具有普通翻译
+        hasLrcTran: false,
+        // 是否具有普通音译
+        hasLrcRoma: false,
+        // 是否具有逐字歌词
         hasYrc: false,
+        // 是否具有逐字翻译
+        hasYrcTran: false,
+        // 是否具有逐字音译
+        hasYrcRoma: false,
+        // 普通歌词数组
+        lrc: [],
+        // 逐字歌词数据
+        yrc: [],
       },
       // 本地歌曲目录
       localSongPath: [],
