@@ -1,3 +1,8 @@
+// BlobUrl
+let lastSongBlobUrl = null;
+let lastCoverBlobUrl = null;
+let lastDownloadBlobUrl = null;
+
 /**
  * 判断当前运行环境
  */
@@ -59,27 +64,15 @@ export const getCacheData = async (key, time, request, params) => {
  */
 export const getLocalCoverData = async (path, isAlbum = false) => {
   try {
-    let blobUrl = null;
+    // 清理过期的 Blob 链接
+    if (lastCoverBlobUrl) URL.revokeObjectURL(lastCoverBlobUrl);
     const coverData = await electron.ipcRenderer.invoke("getMusicCover", path);
     if (coverData) {
       // 将 Uint8Array 数据转换为 Blob
       const blob = new Blob([coverData.coverData], { type: `image/${coverData.coverFormat}` });
       // 生成Blob URL
-      blobUrl = URL.createObjectURL(blob);
-      // 检查当前path是否与上次不一致
-      const previousPath = sessionStorage.getItem("localCoverPath");
-      if (previousPath && previousPath !== path) {
-        // 清除上次的内容
-        const previousBlobUrl = sessionStorage.getItem("localCoverBlobUrl");
-        if (previousBlobUrl) {
-          URL.revokeObjectURL(previousBlobUrl);
-          sessionStorage.removeItem("localCoverBlobUrl");
-        }
-      }
-      // 存储当前path和Blob URL
-      sessionStorage.setItem("localCoverPath", path);
-      sessionStorage.setItem("localCoverBlobUrl", blobUrl);
-      return blobUrl;
+      lastCoverBlobUrl = URL.createObjectURL(blob);
+      return lastCoverBlobUrl;
     } else {
       // 如果没有封面数据
       return `/images/pic/${isAlbum ? "album" : "song"}.jpg?assest`;
@@ -299,35 +292,40 @@ export const generateId = (fileName) => {
  * @param {String} songName - 歌曲名称
  * @returns {number} - 生成的数字ID
  */
-export const downloadFile = async (data, song, path = null) => {
+export const downloadFile = async (data, song, lyric, options) => {
   try {
     const isElectron = checkPlatform.electron();
     const songType = data.type.toLowerCase();
     // 歌曲名称
     const songName =
       song.name +
-      " - " +
+      "-" +
       (Array.isArray(song.artists)
-        ? song.artists.map((ar) => ar.name).join(" / ")
+        ? song.artists.map((ar) => ar.name).join("&")
         : song.artists || "未知歌手");
-    if (isElectron && path) {
-      console.log("开始下载：", data, song, songName, songType, path);
+    if (isElectron && options.path) {
+      console.log("开始下载：", data, song, songName, songType, options.path);
       return await electron.ipcRenderer.invoke(
         "downloadFile",
-        data,
-        JSON.stringify(song),
-        songName,
-        songType,
-        path,
+        JSON.stringify({
+          url: data?.url,
+          data: song,
+          lyric: lyric,
+          name: songName,
+          type: songType,
+        }),
+        JSON.stringify(options),
       );
     } else {
-      const songRes = await fetch(data.url.replace(/^http:/, "https:"));
+      // 清理过期的 Blob 链接
+      if (lastDownloadBlobUrl) URL.revokeObjectURL(lastDownloadBlobUrl);
+      const songRes = await fetch(data?.url.replace(/^http:/, "https:"));
       if (!songRes.ok) throw new Error("下载出错，请重试");
       const blob = await songRes.blob();
-      const url = window.URL.createObjectURL(blob);
+      lastDownloadBlobUrl = URL.createObjectURL(blob);
       // 下载数据
       const a = document.createElement("a");
-      a.href = url;
+      a.href = lastDownloadBlobUrl;
       a.download = `${songName}.${songType}`;
       document.body.appendChild(a);
       a.click();
@@ -359,12 +357,10 @@ export const formatBytes = (bytes, decimals = 2) => {
  * 获取音频文件的 Blob 链接
  * @param {string} url - 音频文件的网络链接
  */
-// 上次生成的 BlobUrl
-let lastBlobUrl = null;
 export const getBlobUrlFromUrl = async (url) => {
   try {
     // 清理过期的 Blob 链接
-    if (lastBlobUrl) URL.revokeObjectURL(lastBlobUrl);
+    if (lastSongBlobUrl) URL.revokeObjectURL(lastSongBlobUrl);
     // 是否为网络链接
     if (!url.startsWith("http://") && !url.startsWith("https://") && !url.startsWith("blob:")) {
       return url;
@@ -377,8 +373,8 @@ export const getBlobUrlFromUrl = async (url) => {
     }
     const blob = await response.blob();
     // 转换为本地 Blob 链接
-    lastBlobUrl = URL.createObjectURL(blob);
-    return lastBlobUrl;
+    lastSongBlobUrl = URL.createObjectURL(blob);
+    return lastSongBlobUrl;
   } catch (error) {
     console.error("获取 Blob 链接遇到错误：" + error);
     throw error;
