@@ -5,7 +5,7 @@
       v-if="allAlbumData"
       :style="{
         height: `calc(100vh - ${
-          Object.keys(music.playSongData)?.length && status.showPlayBar ? 380 : 300
+          Object.keys(music.playSongData)?.length && status.showPlayBar ? 445 : 365
         }px)`,
       }"
       class="local-album"
